requirements.txt
google-api-python-client
google-auth
googleapiclient
gspread
gspread-formatting
gspread_dataframe
oauth2client
pandas
